requirements.txt
fastapi
uvicorn
aiohttp
sqlalchemy
asyncpg
psycopg2-binary
pytest
alembic
python-decouple
requests
pydantic
pydantic-settings
httpx
passlib
python-jose
python-multipart
email_validator
pytz
bcrypt